Adding date navigation
Added a heading link that contains todays date as well as three days in
the past and three days in the future.
Selecting any of the days will re-calculate the date again with 3 days
in the pas and 3 in the future from that date.
Clicking the Habits link will center us back on today.
The links have also been added to the New habit page.